use isl_pw_affs to represent array bounds instead of isl_pw_qpolynomial_folds
The array bounds are (piecewise) affine expression, so an isl_pw_aff
is a better fit.  isl_pw_aff objects are also easier to manipulate.
They weren't used from the start because isl didn't support isl_pw_aff
objects when this code was written.
Another benefit of using isl_pw_affs is that the bound can be computed exactly.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>