Check cuda API and kernel calls for successful execution
Cuda API calls give a return value that specifies if they were executed
sucessfully and a cuda kernel sets a global error buffer, if a problem
occurred. We add runtime checks to check these error conditions.  Before
this patch, we could not even notice, if a kernel was not executed at
all.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>