construct shared memory copying statements inline
Using defines is somewhat fragile as the those defines may end up
referring to parameters that have a fixed value and therefore
do not appear in the generated code.
Since we already have the infrastructure for printing private
memory copying statements, we can simply reuse it for shared memory copying.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>