extract_entire_host_domain: allow some of the user statements to be guarded
When we encounter a user statement in print_stmt, the print_user_stmt_list
function is called (if set) and from within this function, we call
extract_entire_host_domain.  The original user statement may however
occasionally be followed by guarded user statements, so we need to
take this possibility into account.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>