handle absence of parallelism
Before, we would abort if we couldn't find any parallelism, even
if this was just for some of the statements.
We now accept that some parts of the schedule tree do not contain
any parallelism.

After this commit, the sequential statements are still executed
on the GPU (using a single block and a single thread).
It would probably be better to just run the sequential statements
on the host.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>