compute_array_dim_size: detect equalities on index expression
In some cases, the array index may be uniquely determined
by the parameters, while this property may not be apparent
from the constraints.  compute_size_in_direction may then
fail to find the right lower bound and compute an overly
large approximation of the size.
Detecting equalities makes this relation explicit, so
that compute_size_in_direction is able to find the
appropriate lower bound and size (1).

Reported-by: Carlos Juega Reimundez <cjuega@fdi.ucm.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>