use isl_pw_multi_aff to describe array index
Before we would use a list of isl_affs, which is not only more cumbersome,
but also does not allow for a piecewise index expression.
Moreover, the way the list of isl_affs was extracted from the index set,
would not work in all cases.  Now, we simply call isl_pw_multi_aff_from_set
instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>