cuda: allow copy to/from shared memory outside of outermost loop
We already look for the innermost loop affecting the position
of the shared/private memory tile of a particular array and then
copy that array to/from shared memory inside that loop, but if
this position is not affected by any loop, we copying would be
performed inside the innermost loop.
Fix this by also allowing the copying to be performed outside
of any loops.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>