Add support for the generation of c code
C code generation is currently trivial. We just detect the scop and
code generate it with CLoog, but do not apply any further
transformations.  This commit is meant to add the general infrastructure,
but it obviously generates trivially correct code. The c code printing
is enabled by the flag --target=c.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>