only set GPU-specific default options when target is a GPU
In particular, the --schedule-outer-coincidence option is required
for mapping code to a GPU since a schedule band without any outer
coincidence cannot be mapped to blocks and threads.
However, there is no such requirement for code that is mapped to a CPU.
Note that the CPU backend currently does not perform any scheduling,
but this will be changed in an upcoming commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>