cuda.c: interchange_for_unroll: ignore first gen->shared_len loops
These loops have already been generated, so we can't unroll them anymore.
This change does not have any effect on the current code because
these loops have been equated to parameters and therefore don't appear
in any index expressions.  However, this change is required if we remove
the introduction of these parameters.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>