gpu.c: copy_group_shared_accesses: also copy entire strided tiles
The original version would only replace the actual reads by a read
of the entire tile if the tile did not involve any strides.
Change this to also read an entire tile in case of strides.
This has been reported to result in speed-ups on the PolyBench
seidel-2d benchmark.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>