print declarations for variables declared inside the scop
For GPU code generation, we only declare those variables that
are still visible after the scop.  Those that are only visible
inside the scop are only used on the GPU.
For CPU code generation, we need to declare both type of variables,
but make sure that the variables that are not visible after
the scop remain that way.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>