gpu.c: compute_schedule: take domain from ppcg_scop
The original code would extract them from the prog->stmts,
but the domain is already available from the ppcg_scop,
so there is no need to keep track of them in prog->stmts anymore.
Using the domain from the ppcg_scop will allow the gpu code
generator to take advantage of operations performed on this
domain that are shared with the cpu code generator.

Note that this may result in some changes in the generated
code because the order of the statements in the internal
representation may be different.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>