gpu.c: interchange_for_unroll: perform test on isl_pw_multi_aff
interchange_for_unroll, and in particular check_unroll, which is called
from interchange_for_unroll, assumes that the index expression can be
recovered from an equality on the ouput dimension in the map
describing the index expression.  However, this does not work if any
integer divisions are involved.

Explicitly convert the map to an isl_pw_multi_aff and check the affine
functions in the result instead.

Note that we could not do this originally because there were no
isl_pw_multi_aff objects at the time.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>