Prepare polybench_test.sh for multiple options
  * Introduce run_tests method
  * Place generated files into a temporary directory
  * Name generated files descriptively
  * Use flag --std=gnu99
  * Keep files until the end of the test run

Enforcing gnu99 is required to compile the ppcg generated code with gcc.

Keeping the files until the end of the test run allows the inspection of
the generated files, in case incorrect code is generated and the test
is aborted. The overall size of generated files is less than 10 Mb, so
this should not cause any problems.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>