gpu.c: compute private tile during array reference grouping
The private tile may be loaded at an outer loop and may therefore
have to be merged with other groups based on the loop where the loading
is performed.  We therefore need to compute the private tile
during the array reference grouping.  The additional merging will
be handled in a subsequent commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>