copy text from input file based on file offsets rather than location of pragmas
If the --pet-autodetect option is specified, then there may be no pragmas
in the input file and the transformed code could get inserted at the
end or at some other incorrect position.

Note that if a scop is detected based on the scop and endscop pragmas
(i.e., if the --pet-autodetect option is not specified), then
the offsets of the scop as returned by pet include the pragmas.
Since PPCG replaces the entire scop, these pragmas will no longer appear
in the output.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>