cuda: Fix self assignment in cudaCheckKernel macro
Passing a variable to a macro that has the same name as a
variable within the macro can cause surprising behavior. In our
case we created a self assignment 'e = (e)', which references the
new variable instead of the parameter that was passed into the
function. We fix this by including the macro name in the variables used
within the macro. This should reduce the probability of name clashes.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>