gpu.c: use at_each_node to attach copy statement information
The original code used a create_leaf callback to create an isl_ast_node
representing the copy statement with copy statement attached to it.
However, create_leaf callbacks are expected to perform a nested AST
generation as the schedule passed to the callback may still contain
constraints that have not yet been expressed in the AST generated so far.
Since we essentiall ignore the schedule constraints in the callback,
in some cases this would result in constraints disappearing.

By using an at_each_domain callback instead, we avoid this problem
as isl will generate all necessary constraints.  We do not even
need to create an isl_ast_node since isl will already have created
it for us.  The exact form of this node may be a bit different
from the one that we used to create in the create_leaf callback,
but this shouldn't have any impact since we only use the information
attached to the node during printing rather than the node expression itself.

Note that create_sync_leaf performs a similar operation to what
the create_copy_leaf callback used to do, but this should be harmless
as even if some constraints end up getting dropped we would at most perform
a few extra synchronizations.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>