gpu.c: compute effective block size
The default block sizes or those specified by the user may result in
idle threads based on other options such as the tile sizes and the
values of the parameters.
The effective block sizes are computed after tiling and take into
account the parameter values.  This should result in fewer idle threads.
The effective block sizes are used in the definition of the block and
to bound the thread ids.  The code for copying data to/from shared
memory also takes into account the effective block sizes.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>