gpu.c: copy_access: only encode access type in iteration domain name
The original code would encode both the access type and the local
array name in the iteration domain name, but only the access type
was being used later on.  The local array name is also available
in the nested space.
Making the name independent of the local array name makes it
easier to refactor the code that sets this local array name.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>