gpu.c: mark_outer_tilable: split tilable band to match tile length
If the number of specified tile sizes for the current kernel
is smaller than the dimension of the band node, then we need to split off
a piece that has the appropriate dimension.
This change does not have any effect at the moment since
the band and its descendants are currently replaced by a mark node,
but it is required to be able to tile the band node.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>