update pet for change in pet_scop_collect_{reads,writes}
These now support the collection of data dependent accesses.
Since we currently do not support these in PPCG, we simply
collect all "may" accesses, knowing that they are equal to the
"must" accesses.

Since the ranges of the access relations are now also intersected
with the array extents by pet, also add a note to README stating
that -DPOLYBENCH_USE_C99_PROTO should be used when processing
PolyBench.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>