add support for arrays of structs
pet takes care of parsing arrays of structs and together with isl,
it also takes care of printing accesses to arrays of structs.
Since we assume that there is no aliasing, dependence analysis
does not need to be modified.
Elements of the outer arrays of structs are transferred as a whole
to the device, which means we need to map the member accesses
to the corresponding element of the outer array during some
parts of the processing.
We currently do not allow arrays of structs to be mapped to
shared or private memory.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>