gpu.c: check_shared_memory_bound: ignore groups that are mapped to registers
For a given array reference group, we may have computed both a tile
for mapping to shared memory and a tile for mapping to registers.
If so, the group will be mapped to registers and we should ignore the
size of the tile for mapping to shared memory when computing
the total amount of shared memory used.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>