gpu: avoid copying in data from local arrays
If an array that is local to the scop is read without first being written,
then there is no point in copying it in since it cannot have been
written prior to the scop.  Local arrays are not even declared in the host
code, so it is even impossible to perform the copying.
Warn about the uninitialized read instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>