gpu.c: create_kernel: obtain number of block dimensions from marked band
The initial number of block dimensions is currently always equal
to the initial number of grid dimensions because the bands
mapped to blocks and threads are derived from the same band
through tiling.  In future, however, we may want to handle
other cases, so from here on we do not assume any specific
relation between these two bands.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>