gpu: mimic wrap option in schedule tree
Or rather, mimic --no-wrap in the schedule tree.
If the wrap option is not set then in the flat schedule representation
the schedule dimensions that are mapped to block/thread identifiers
are first tiled by the grid/block sizes and then the tile dimensions
are optionally scaled by those sizes again.
For backward compatibility, we need to mimic this behavior in the schedule
tree representation.  If the wrap is not set, we therefore need to
perform the same changes to the bands that are mapped to block/thread
identifiers.

If the wrap option is set, then the original schedule dimensions
are left untouched and this is also what happens in the schedule
tree representation.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>