gpu_group_references: isolate all required schedule information in structure
In a subsequent commit, we will perform the array reference grouping
before the AST generation, which we will need to extract schedule
information from the schedule tree rather than from gpu_gen.
Store all information we need from the gpu_gen in a separate
gpu_group_data data structure to minimize the required changes when
we perform the switch.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>