gpu.c: set_schedule_modulo: allow smaller number of band members than identifiers
If this number is smaller, then the first elements in the list
of identifiers are equated to zero.
We will need this when we will generate the entire code from
a single schedule tree.  In particular, the copying to/from
shared memory will be mapped to the last thread identifiers
rather than the first.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>