ppcg_kernel: initialize array field as soon as ppcg_kernel is created
We currently only store localized copies of the array sizes in
the gpu_local_array_info elements, but we will be adding more
kernel specific information to this data structure.
This information may be set before localize_bounds, so we need
to construct the array of gpu_local_array_info structures beforehand.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>