gpu_group_references: distinguish between shared_schedule and copy_schedule
The initial mapping to shared memory is currently computed right outside
the band that is mapped to threads.  In a subsequent commit, it will
be made possible to introduce this mapping at a higher level.
In preparation, adjust gpu_group_references to take into account
this distinction by introducing a copy_schedule that is used for
determining the initial overlap between accesses.
In particular, when a reference group can only be mapped to private
memory, overlap will still be determined at the deepest level.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>