gpu: move array reference groups from gpu_array_info to gpu_local_array_info
The gpu_array_info elements are shared by the entire gpu_prog,
while the reference groups are kernel specific.
We currently use the reference groups of a kernel before
we consider the next kernel, overwriting the old reference groups.
In the future, however, we will be constructing a single schedule tree
for the entire program, containing information about all kernels
together.  We therefore need to put this information in kernel
specific data structures.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>