gpu_group_references: take into account "shared" mark
The "shared" mark indicates where the mapping to shared memory should
be computed.  The mark is inserted by create_kernel right outside
of the "kernel" mark if it is not already present in the input.
This ensures that the behavior of create_kernel is not affected
if the input does not contain a "shared" mark.

This will be useful for hybrid tiling, where the mapping to shared memory
should be computed outside the inner time dimension.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>