gpu: avoid private memory if unrolling is needed and any_force_private is set
That is, do not even construct a private tile in this case.
Before we would construct such a private tile but then discard it afterwards
inside interchange_for_unroll.  This original check is however performed
after all the array reference grouping has finished.
If both a mapping to shared and private memory is possible then this
grouping would take into account the private tile, while in the end
the shared tile would get used.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>