gpu: stop setting group_coscheduled AST build option on host code
We will be generating both the host and the device code
from a single schedule tree in a subsequent commit.
In that context, the entire schedule tree is available during
AST generation, so there is no point in setting the group_coscheduled
option and there will be no way to obtain the same effect of
the group_coscheduled option in this new context.
Since we are grouping the statements in a kernel there
is already no need to set this option in the current state.
We therefore drop this option during the generation of the host AST first
so that the effect of dropping the option is available separately.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>