gpu: group statement instances that will form a kernel
We currently create kernels during the AST generation, in particular
inside the leaves of the host schedule.
In the future, we will create them in advance and we therefore need
to make sure that the statements that we consider to form a kernel
during the analysis actually end up in a single kernel.
We therefore group the statements that will form a kernel,
provided there is more than one statement.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>