gpu: extract effective block size from schedule tree
Before, we would extract the effective block size based on the domain
of the schedule generated by the AST generator for the host code.
Since we want to insert a context node in the schedule tree expressing
the bounds on the thread identifiers,
we need to extract this set from the schedule tree instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>