gpu: also create (a single) array reference group for arrays of structs
We currently do not map such arrays to shared or private memory,
so there is no real need to compute array reference groups.
However, the fact that there are no array reference groups
is also used as an indication that the array is not accessed at all
and we need to take into account that we cannot draw this conclusion
for arrays of structs.

Create a single combined array reference group for arrays of structs,
assuming the array is accessed at all, so that we can remove the
special case.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>