gpu: avoid mapping independent non-permutable bands to the device
If parts of the input code are completely independent of other parts
then they will end up as children of an outer set node.
If only some of these children have permutable bands, then there
is no point in also mapping the other children to the device.
We therefore separate the children without permutable bands and
keep them on the host.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>