gpu_stmt_access: drop group field
The gpu_stmt_access objects are shared by the entire gpu_prog,
while the reference groups are kernel specific.
We currently use the reference groups of a kernel before
we consider the next kernel, overwriting the old reference groups.
In the future, however, we will be constructing a single schedule tree
for the entire program, containing information about all kernels
together.  We therefore cannot store this information in global
data structures.  In principle, we could fill in this information
right before we need it for a specific kernel.  For now, just look
for the reference group that contains the reference.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>