ppcg_scop: represent tagger as isl_union_pw_multi_aff instead of isl_union_map
In particular, the tagger is now represented as an isl_union_pw_multi_aff
that projects out the tags rather than an isl_union_map that add the tags.
The tags are then added by computing the preimage over the new tagger
rather than applying the old tagger.

Representing the tagger as an isl_union_pw_multi_aff will be useful
for adding tags to a schedule in schedule tree representation since
it is possible to compute the pullback of a schedule tree, but there is
no functionality for applying a union map to the schedule tree domain.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>