gpu.c: compute_schedule: exploit independences
We are still very conservative at this point.
In particular, we only exploit independences when live range reordering
is enabled and we only exploit them to reduce the set of coincidence
constraints.  That means that we do not allow more freedom in the
schedule generation, but only allow some loops in the generated schedule
to be considered parallel that may not have been considere parallel
with the original set of coincidence constraints.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>