do not map sliced accesses to shared/private memory
When we add support for calling user defined functions, we may
run across cases where an entire array or a slice of an array
is passed to a function.
In these cases, we need to make sure that the accessed array
is not mapped to shared or private memory as the called function
may expect the original layout in global memory.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>