opencl backend: opencl_open_files: use base name in #include
The name of the header file may have been derived from the output
file name specified by the user and may therefore include a relative path.
This path is relative to the location where ppcg is invoked
but this location may be different from the one where the generated
code is compiled.  Since the header file is created in the same
directory as the host source file, we can use the base name instead.

This commit fixes the --keep option of polybench_test.sh,
which has been broken ever since the introduction of the OpenCL backend
in 2b8a598 (Add OpenCL backend, Fri Jun 21 11:57:05 2013 +0200).

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>