gpu_local_array_info: keep track of the total number of indices
When passing rows or bigger slices of an array to a function,
the number of indices in an index expression may be smaller than
the total number of indices of an array (i.e., its dimension).
In the next commit we will adjust the linearization to take
this possibility into account, but in order to detect the situation,
we need to keep track of the total number of indices.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>