gpu_stmt_access: represent access to outer array element
We curently do not perform any remapping of arrays that involve
member elements, but the arrays that we consider themselves
are always the outer arrays, so it is more consistent for gpu_stmt_access
to also refer to these outer arrays.
This change also simplifies the next commit as in makes sure
that the entire access relation is mapped onto a single space.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>