gpu: add nodes for copying arrays to the schedule tree
We currently map the entire scop to the device.  In the future
we will want to be able to only map a subtree of the schedule
tree to the device.  This means that the code for copying
the arrays to and from the device needs to be inserted around
that subtree.  Add appropriate nodes to the schedule tree
to ensure to obtain this effect.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>