gpu: compute copy_in and copy_out from schedule tree
We currently map the entire scop to the device.  In the future
we will want to be able to only map a subtree of the schedule
tree to the device.  This means that the copy_in and copy_out
sets should be computed based on this subtree rather than
on the entire scop.
We start by computing these sets from the schedule tree,
even though it is currently passed the entire tree.
We do however already take into account that in the future
this function will also be used on subtrees of the schedule tree.
In particular, we not only need to take into account the data
that may persist after the entire scop, but also the data that
may persist after the subtree inside the scop.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>